--- a/infra/samhstn/buildspec.yml
+++ b/infra/samhstn/buildspec.yml
@@ -1,13 +1,20 @@
 version: 0.2
 
+env:
+  shell: bash
+
 phases:
   install:
     commands:
+      - |
+        if aws s3 ls s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache &> /dev/null;then
+          aws s3 cp --recursive s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/_build _build
+          aws s3 cp --recursive s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/deps deps
+        else
+          aws s3 cp --recursive s3://$MASTER_BUILD_CACHE_BUCKET_NAME/_build _build
+          aws s3 cp --recursive s3://$MASTER_BUILD_CACHE_BUCKET_NAME/deps deps
+        fi
       - mix deps.get
-      - ls -alL _build
-      - ls -alL _build/test/lib/phoenix/
-      - ls -alL _build/test/lib/phoenix/ebin
-      - # cat _build/test/lib/phoenix/ebin/Elixir.Phoenix.ActionClauseError.beam
   pre_build:
     commands:
       - MIX_ENV=test mix compile
@@ -17,7 +24,6 @@ phases:
     commands:
       - mix phx.digest
       - MIX_ENV=prod mix release
-      - echo $S3_CODEBUILD_BUCKET_NAME
       - zip -r ${CODEBUILD_RESOLVED_SOURCE_VERSION}.zip priv
       - zip -qr ${CODEBUILD_RESOLVED_SOURCE_VERSION}.zip _build/prod/rel/samhstn
       - zip -j ${CODEBUILD_RESOLVED_SOURCE_VERSION}.zip
@@ -47,7 +53,6 @@ phases:
       - aws configure --profile root set aws_secret_access_key $(echo $ASSUME_ROLE | jp -u 'Credentials.SecretAccessKey')
       - aws configure --profile root set aws_session_token $(echo $ASSUME_ROLE | jp -u 'Credentials.SessionToken')
       - |
-        #!/bin/bash
         set -e
 
         cat <<EOF > change-batch.json
@@ -73,10 +78,5 @@ phases:
           --profile root \
           --hosted-zone-id $SAMHSTN_HOSTED_ZONE_ID \
           --change-batch file://change-batch.json
-
-cache:
-  paths:
-    - '_build/test/lib/phoenix/**/*'
-    - _build/test/lib/phoenix
-    - _build/test/lib/phoenix/ebin
-    - _build/test/lib/phoenix/ebin/*
+      - aws s3 sync --delete _build s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/_build
+      - aws s3 sync --delete deps s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/deps