--- a/infra/samhstn/buildspec.yml
+++ b/infra/samhstn/buildspec.yml
@@ -8,11 +8,13 @@ phases:
     commands:
       - |
         if aws s3 ls s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache &> /dev/null;then
-          aws s3 cp --recursive s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/_build _build
-          aws s3 cp --recursive s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/deps deps
+          echo "Using branch specific cache."
+          aws s3 cp --quiet --recursive s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/_build _build
+          aws s3 cp --quiet --recursive s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/deps deps
         else
-          aws s3 cp --recursive s3://$MASTER_BUILD_CACHE_BUCKET_NAME/_build _build
-          aws s3 cp --recursive s3://$MASTER_BUILD_CACHE_BUCKET_NAME/deps deps
+          echo "Branch specific cache doesn't exist, using cache from master if it exists."
+          aws s3 cp --quiet --recursive s3://$MASTER_BUILD_CACHE_BUCKET_NAME/_build _build || true
+          aws s3 cp --quiet --recursive s3://$MASTER_BUILD_CACHE_BUCKET_NAME/deps deps || true
         fi
       - mix deps.get
   pre_build:
@@ -23,7 +25,7 @@ phases:
       - mix test
   build:
     commands:
-      - mix phx.digest
+      - MIX_ENV=prod mix phx.digest
       - MIX_ENV=prod mix release
       - zip -r ${CODEBUILD_RESOLVED_SOURCE_VERSION}.zip priv
       - zip -qr ${CODEBUILD_RESOLVED_SOURCE_VERSION}.zip _build/prod/rel/samhstn
@@ -79,5 +81,6 @@ phases:
           --profile root \
           --hosted-zone-id $SAMHSTN_HOSTED_ZONE_ID \
           --change-batch file://change-batch.json
-      - aws s3 sync --delete _build s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/_build
-      - aws s3 sync --delete deps s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/deps
+      - aws s3 sync --quiet --delete _build/prod/lib s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/_build/prod/lib
+      - aws s3 sync --quiet --delete _build/test s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/_build/test
+      - aws s3 sync --quiet --delete deps s3://${S3_CODEBUILD_BUCKET_NAME}/${ISSUE_NUMBER}/cache/deps