--- a/infra/samhstn/deletespec.yml
+++ b/infra/samhstn/deletespec.yml
@@ -1,5 +1,8 @@
 version: 0.2
 
+env:
+  shell: bash
+
 phases:
   build:
     commands:
@@ -15,7 +18,6 @@ phases:
       - aws configure --profile root set aws_secret_access_key $(echo $ASSUME_ROLE | jp -u 'Credentials.SecretAccessKey')
       - aws configure --profile root set aws_session_token $(echo $ASSUME_ROLE | jp -u 'Credentials.SessionToken')
       - |
-        #!/bin/bash
         set -e
 
         cat <<EOF > change-batch.json