Fixes #98

Before cache, total: **214 secs**

```
SUBMITTED	<1 sec	
QUEUED	<1 sec	
PROVISIONING	60 secs	
DOWNLOAD_SOURCE	2 secs	
INSTALL	1 sec	
PRE_BUILD	39 secs	
BUILD	41 secs	
POST_BUILD	69 secs	
UPLOAD_ARTIFACTS	<1 sec	
FINALIZING	2 secs
```

After cache - initial build, total: **260 secs**

```
SUBMITTED	<1 sec
QUEUED	1 sec
PROVISIONING	75 secs
DOWNLOAD_SOURCE	2 secs
INSTALL	6 secs
PRE_BUILD	42 secs
BUILD	44 secs
POST_BUILD	88 secs
UPLOAD_ARTIFACTS	<1 sec
FINALIZING	2 secs
```

After cache - subsequent builds, total: **106 secs**

```
SUBMITTED	<1 sec
QUEUED	1 sec
PROVISIONING	75 secs
DOWNLOAD_SOURCE	2 secs
INSTALL	14 secs
PRE_BUILD	3 secs
BUILD	3 secs
POST_BUILD	6 secs
UPLOAD_ARTIFACTS	<1 sec
FINALIZING	2 secs
```

This is the initial difference in build times, and there will likely be ways to make it faster.

Caching like this makes it possible to use `dialyzer` (without waiting for 5 minute plt file builds each time) (see #94)